+++ src/modules/Playerbot/Group/PlayerbotGroupScript.cpp
@@ -396,10 +396,10 @@ void PlayerbotGroupScript::InitializeGroupState(Group* group, GroupState& state)
 
     // NOTE: Target icon initialization removed - Group::GetTargetIcon() is not accessible
 
-    // Initialize difficulty
-    state.dungeonDifficulty = static_cast<uint8>(group->GetDungeonDifficultyID());
-    state.raidDifficulty = static_cast<uint8>(group->GetRaidDifficultyID());
-    state.legacyRaidDifficulty = static_cast<uint8>(group->GetLegacyRaidDifficultyID());
+    // Initialize difficulty (WoW 12.0: Difficulty is int16)
+    state.dungeonDifficulty = static_cast<int16>(group->GetDungeonDifficultyID());
+    state.raidDifficulty = static_cast<int16>(group->GetRaidDifficultyID());
+    state.legacyRaidDifficulty = static_cast<int16>(group->GetLegacyRaidDifficultyID());
 
     // Initialize raid status
     state.isRaid = group->isRaidGroup();
